\name{rcpptest-package}
\alias{rcpptest-package}
\alias{rcpptest}
\docType{package}
\title{\packageTitle{rcpptest}}
\description{\packageDescription{rcpptest}}
\section{Package Content}{\packageIndices{rcpptest}}
\author{\packageAuthor{rcpptest}}
\section{Maintainer}{\packageMaintainer{rcpptest}}
\keyword{package}
