% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpptest.R
\docType{package}
\name{rcpptest}
\alias{rcpptest}
\alias{rcpptest-package}
\title{rqPen: A package for estimating quantile regression models using penalized objective functions.}
\description{
The package estimates a quantile regression model using LASSO, Adaptive LASSO, SCAD, MCP, elastic net, 
and their group counterparts, with the exception of elastic net for which there is no group penalty implementation.
}
