% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{some_math}
\alias{some_math}
\title{Times Two Plus 5}
\usage{
some_math(x)
}
\arguments{
\item{x}{A number to multiply by 2 and add 5 to}
}
\description{
Multiplies an number by 2, then adds 5.
}
\examples{
some_math(5)
some_math(3)
some_math(10)
}
